# [Tokenizer.cs] String.Equals instead of String.Compare gives 3x speedup